# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E_Messe API",
        "description": "API pour l'application E_Messe - Gestion des messes",
        "contact": {
            "email": "leprodev03@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081/api",
            "description": "Serveur Local"
        },
        {
            "url": "https://votre-domaine.com/api",
            "description": "Serveur de Production"
        }
    ],
    "paths": {
...
        "/api/favoris": {
            "get": {
                "tags": [
                    "Favoris"
                ],
                "summary": "Liste des favoris de l'utilisateur connecté",
                "description": "Récupère la liste complète des paroisses que l'utilisateur a ajoutées à ses favoris.",
                "operationId": "1445c8b652df3c8c49aee1bd1a896298",
                "responses": {
                    "200": {
                        "description": "Liste des favoris récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "paroisse_id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "paroisse": {
                                                        "properties": {
                                                            "nom": {
                                                                "type": "string",
                                                                "example": "Paroisse Saint Michel"
                                                            },
                                                            "commune": {
                                                                "properties": {
                                                                    "nom": {
                                                                        "type": "string",
                                                                        "example": "Yopougon"
                                                                    },
                                                                    "ville": {
                                                                        "properties": {
                                                                            "nom": {
                                                                                "type": "string",
                                                                                "example": "Abidjan"
                                                                            }
                                                                        },
                                                                        "type": "object"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Utilisateur non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Favoris"
                ],
                "summary": "Ajouter une paroisse aux favoris",
                "description": "Permet d'ajouter une paroisse à la liste des favoris de l'utilisateur connecté.",
                "operationId": "67a41e7b4a7468dfa0f366c958e2b38f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "paroisse_id"
                                ],
                                "properties": {
                                    "paroisse_id": {
                                        "description": "ID de la paroisse à ajouter",
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paroisse ajoutée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Paroisse ajoutée aux favoris"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "paroisse_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Utilisateur non authentifié"
                    },
                    "422": {
                        "description": "Validation échouée (paroisse inexistante)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/favoris/{id}": {
            "delete": {
                "tags": [
                    "Favoris"
                ],
                "summary": "Supprimer un favori",
                "description": "Supprime une paroisse des favoris de l'utilisateur connecté.",
                "operationId": "160a3a22b01929f8288447de43f9e3a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du favori à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favori supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Favori supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Favori introuvable"
                    },
                    "401": {
                        "description": "Utilisateur non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}